Fixed the select all bug with company and product as well as the drop down width issue. Added a simple all selection in the dropdowns.
I noticed that if you select two companies and an issue, the timeframe may be limited under the hood to only the time frame for the company-product cross section with the shorter time frame. For example, selecting Equifax and another company as well as crediting reporting (the long option) for the product, will cut out several years of data for Equifax.